+++ evm/Cargo.toml
@@ -44,8 +44,6 @@ proptest = "1.0.0"
 # Display
 yansi = "0.5.1"
 url = "2.2.2"
-tui = { version = "0.16.0", default-features = false, features = ["crossterm"] }
-
 
 [dev-dependencies]
 tempfile = "3.3.0"
