+++ evm/src/debug.rs
@@ -124,6 +124,8 @@ pub struct DebugStep {
     pub ic: usize,
     /// Cumulative gas usage
     pub total_gas_used: u64,
+    /// Associated max seen trace index
+    pub max_trace_index: usize,
     /// Associated trace index
     pub trace_index: usize,
 }
@@ -138,6 +140,7 @@ impl Default for DebugStep {
             push_bytes: None,
             ic: 0,
             total_gas_used: 0,
+            max_trace_index: 0,
             trace_index: 0,
         }
     }
