+++ evm/src/executor/inspector/debugger.rs
@@ -47,7 +47,10 @@ pub struct Debugger {
     pub previous_gas_block: u64,
     /// Call Trace tracker
     pub call_trace_tracker: Rc<RefCell<Vec<usize>>>,
+    /// Max trace index seen
     pub max_trace_index_seen: usize,
+    /// Active trace index
+    pub active_trace_index: usize,
 }
 
 impl Debugger {
@@ -78,9 +81,9 @@ impl Debugger {
 
     /// Enters a new execution context.
     pub fn enter(&mut self, depth: usize, address: Address, kind: CallKind) {
-        let curr_trace_index = *(self.call_trace_tracker.borrow().iter().last().unwrap_or(&0));
-        if curr_trace_index > self.max_trace_index_seen {
-            self.max_trace_index_seen = curr_trace_index;
+        self.active_trace_index = *(self.call_trace_tracker.borrow().iter().last().unwrap_or(&0));
+        if self.active_trace_index > self.max_trace_index_seen {
+            self.max_trace_index_seen = self.active_trace_index;
         }
 
         self.context = address;
@@ -89,6 +92,7 @@ impl Debugger {
 
     /// Exits the current execution context, replacing it with the previous one.
     pub fn exit(&mut self) {
+        self.active_trace_index = *(self.call_trace_tracker.borrow().iter().last().unwrap_or(&0));
         if let Some(parent_id) = self.arena.arena[self.head].parent {
             let DebugNode { depth, address, kind, .. } = self.arena.arena[parent_id];
             self.context = address;
@@ -119,7 +123,8 @@ where
                 instruction: Instruction::Cheatcode(
                     call.input[0..4].try_into().expect("malformed cheatcode call"),
                 ),
-                trace_index: self.max_trace_index_seen,
+                max_trace_index: self.max_trace_index_seen,
+                trace_index: self.active_trace_index,
                 ..Default::default()
             });
         }
@@ -191,7 +196,8 @@ where
                 .get(&pc)
                 .expect("unknown ic for pc"),
             total_gas_used: gas_used(data.env.cfg.spec_id, total_gas_spent, gas.refunded() as u64),
-            trace_index: self.max_trace_index_seen,
+            max_trace_index: self.max_trace_index_seen,
+            trace_index: self.active_trace_index,
         });
 
         Return::Continue
