+++ ui/Cargo.toml
@@ -15,3 +15,4 @@ hex = "0.4.3"
 ethers = { git = "https://github.com/gakonst/ethers-rs" }
 forge = { path = "../forge" }
 revm = { version="1.3", default-features = false, features = ["std", "k256", "with-serde"] }
+foundry-evm = { path = "../evm" }
